requirements.txt
pandas
numpy
matplotlib
seaborn
websocket-client 